+++ ByteDance/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 6.5.0.8.0
+* Certified with ByteDance SDK 6.5.0.8.
+* Removed deprecated code paths based on the minimum supported AppLovin MAX SDK version 13.2.0.
+* Updated minimum Android API level to 19 to match ByteDance SDK.
+
 ## 6.5.0.6.0
 * Certified with ByteDance SDK 6.5.0.6.